refactor: rewrite simple_udt by TypeScript
